move constants and class methods to their own file
